converted  IEXSharp.Model.CoreData.StockPrices.Response.updated from long? to decimal?
- fixes #116

+++ IEXSharp/Model/CoreData/StockPrices/Response/HistoricalPriceResponse.cs
@@ -19,7 +19,7 @@ namespace IEXSharp.Model.CoreData.StockPrices.Response
 		public string id { get; set; }
 		public string key { get; set; }
 		public string subkey { get; set; }
-		public long? updated { get; set; }
+		public decimal? updated { get; set; }
 		public decimal? changeOverTime { get; set; }
 		public decimal? marketChangeOverTime { get; set; }
 		public decimal? uOpen { get; set; }
